🎨 renamed yahoo tool data fetch folder
